Adding wizard for send flow
