+++ app/components/Views/confirmations/components/send/amount/amount-edit/amount-edit.styles.ts
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native';
+
+import { Theme } from '../../../../../../../util/theme/models';
+
+export const styleSheet = (params: { theme: Theme }) => {
+  const { theme } = params;
+  return StyleSheet.create({
+    input: {
+      borderWidth: 1,
+      borderColor: theme.colors.border.default,
+    },
+  });
+};
