+++ app/components/Views/confirmations/components/send/amount/amount-edit/index.ts
@@ -1,0 +1,1 @@
+export { AmountEdit } from './amount-edit';
