+++ app/components/Views/confirmations/components/send/amount/amount.styles.ts
@@ -1,13 +1,8 @@
 import { StyleSheet } from 'react-native';
 
-import { Theme } from '../../../../../../util/theme/models';
-
-export const styleSheet = (params: { theme: Theme }) => {
-  const { theme } = params;
-  return StyleSheet.create({
-    input: {
-      borderWidth: 1,
-      borderColor: theme.colors.border.default,
+export const styleSheet = () =>
+  StyleSheet.create({
+    container: {
+      padding: 20,
     },
   });
-};
