+++ app/components/Views/confirmations/components/send/amount/amount.test.tsx
@@ -1,111 +1,236 @@
 import React from 'react';
-
-import renderWithProvider from '../../../../../../util/test/renderWithProvider';
-// eslint-disable-next-line import/no-namespace
-import * as MaxAmountUtils from '../../../hooks/send/useMaxAmount';
-// eslint-disable-next-line import/no-namespace
-import * as ConversionUtils from '../../../hooks/send/useConversions';
-import { SendContextProvider } from '../../../context/send-context';
-import { evmSendStateMock } from '../../../__mocks__/send.mock';
-import { Amount } from './amount';
+import { ParamListBase, RouteProp, useRoute } from '@react-navigation/native';
 import { fireEvent } from '@testing-library/react-native';
+import { merge } from 'lodash';
+
+import renderWithProvider, {
+  ProviderValues,
+} from '../../../../../../util/test/renderWithProvider';
+import { SendContextProvider } from '../../../context/send-context';
+import {
+  ACCOUNT_ADDRESS_MOCK_1,
+  TOKEN_ADDRESS_MOCK_1,
+  evmSendStateMock,
+} from '../../../__mocks__/send.mock';
+import { Amount } from './amount';
+
+jest.mock('../../../../../../core/Engine', () => ({
+  context: {
+    NetworkController: {
+      findNetworkClientIdByChainId: jest.fn().mockReturnValue('mainnet'),
+    },
+    AssetsContractController: {
+      getERC721AssetSymbol: Promise.resolve(undefined),
+    },
+  },
+}));
 
 jest.mock(
   '../../../../../../components/Views/confirmations/hooks/gas/useGasFeeEstimates',
   () => ({
-    useGasFeeEstimates: () => ({ gasFeeEstimates: {} }),
+    useGasFeeEstimates: () => ({
+      gasFeeEstimates: { medium: { suggestedMaxFeePerGas: 1.5 } },
+    }),
   }),
 );
 
+const mockGoBack = jest.fn();
+const mockNavigate = jest.fn();
 jest.mock('@react-navigation/native', () => ({
   ...jest.requireActual('@react-navigation/native'),
   useNavigation: () => ({
-    goBack: jest.fn(),
-    navigate: jest.fn(),
+    goBack: mockGoBack,
+    navigate: mockNavigate,
   }),
   useRoute: jest.fn().mockReturnValue({
     params: {
       asset: {
         chainId: '0x1',
-        name: 'Ethereum',
-        address: '0x123',
+        address: '0x935E73EDb9fF52E23BaC7F7e043A1ecD06d05477',
       },
     },
   }),
 }));
 
-const renderComponent = () =>
-  renderWithProvider(
+const renderComponent = (mockState?: ProviderValues['state']) => {
+  const state = mockState
+    ? merge(evmSendStateMock, mockState)
+    : evmSendStateMock;
+  return renderWithProvider(
     <SendContextProvider>
       <Amount />
     </SendContextProvider>,
     {
-      state: evmSendStateMock,
+      state,
     },
   );
+};
 
 describe('Amount', () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
+
   it('renders correctly', async () => {
     const { getByText } = renderComponent();
 
+    expect(getByText('From:')).toBeTruthy();
     expect(getByText('Value:')).toBeTruthy();
   });
 
-  it('display option to set value to Max', async () => {
+  it('use from address returned from SendContext', async () => {
     const { getByText } = renderComponent();
 
-    expect(getByText('Max')).toBeTruthy();
+    expect(getByText('From:')).toBeTruthy();
+    expect(getByText(ACCOUNT_ADDRESS_MOCK_1)).toBeTruthy();
   });
 
-  it('does not display Max option if it is not supported', async () => {
-    jest.spyOn(MaxAmountUtils, 'useMaxAmount').mockReturnValue({
-      getMaxAmount: () => undefined,
-      isMaxAmountSupported: false,
-    });
-    const { queryByText } = renderComponent();
-
-    expect(queryByText('Max')).toBeNull();
+  it('display error for invalid amount value', async () => {
+    const { getByText, getByTestId } = renderComponent();
+    fireEvent.changeText(getByTestId('send_amount'), 'abc');
+    expect(getByText('Invalid amount')).toBeTruthy();
   });
 
-  it('update amount with max value when max button is clicked', async () => {
-    const MAX_AMOUNT = '0.01234';
-    jest.spyOn(MaxAmountUtils, 'useMaxAmount').mockReturnValue({
-      getMaxAmount: () => MAX_AMOUNT,
-      isMaxAmountSupported: true,
-    });
-    const { getByTestId, getByText } = renderComponent();
+  it('display error if amount is greater than balance for native token', async () => {
+    (useRoute as jest.MockedFn<typeof useRoute>).mockReturnValue({
+      params: {
+        asset: {
+          isNative: true,
+          chainId: '0x1',
+        },
+      },
+    } as RouteProp<ParamListBase, string>);
+
+    const { getByText, getByTestId } = renderComponent();
+    fireEvent.changeText(getByTestId('send_amount'), '100');
+    expect(getByText('Insufficient funds')).toBeTruthy();
+  });
+
+  it('display error if amount is greater than balance for ERC20 token', async () => {
+    (useRoute as jest.MockedFn<typeof useRoute>).mockReturnValue({
+      params: {
+        asset: {
+          address: TOKEN_ADDRESS_MOCK_1,
+          decimals: 2,
+        },
+      },
+    } as RouteProp<ParamListBase, string>);
+
+    const { getByText, getByTestId } = renderComponent();
+    fireEvent.changeText(getByTestId('send_amount'), '100');
+    expect(getByText('Insufficient funds')).toBeTruthy();
+  });
+
+  it('asset passed in nav params should be used if present', () => {
+    (useRoute as jest.MockedFn<typeof useRoute>).mockReturnValue({
+      params: {
+        asset: {
+          name: 'Ethereum',
+          address: TOKEN_ADDRESS_MOCK_1,
+        },
+      },
+    } as RouteProp<ParamListBase, string>);
+    const { getByText } = renderComponent();
+    expect(getByText(`Asset: ${TOKEN_ADDRESS_MOCK_1}`)).toBeTruthy();
+  });
+
+  it('pressing Max uses max balance of ERC20 token', () => {
+    (useRoute as jest.MockedFn<typeof useRoute>).mockReturnValue({
+      params: {
+        asset: {
+          address: TOKEN_ADDRESS_MOCK_1,
+          decimals: 2,
+        },
+      },
+    } as RouteProp<ParamListBase, string>);
+
+    const { getByText, getByTestId } = renderComponent();
+    expect(getByTestId('send_amount').props.value).toBe('');
     fireEvent.press(getByText('Max'));
-    expect(getByTestId('send_amount').props.value).toBe(MAX_AMOUNT);
+    expect(getByTestId('send_amount').props.value).toBe('0.05');
+    expect(getByText('$ 0.05')).toBeTruthy();
   });
 
-  it('display option for fiat toggle', async () => {
-    const { getByTestId } = renderComponent();
+  it('pressing Max uses max balance minus gas for native token', () => {
+    (useRoute as jest.MockedFn<typeof useRoute>).mockReturnValue({
+      params: {
+        asset: {
+          isNative: true,
+          chainId: '0x1',
+          address: TOKEN_ADDRESS_MOCK_1,
+        },
+      },
+    } as RouteProp<ParamListBase, string>);
 
-    expect(getByTestId('fiat_toggle')).toBeTruthy();
+    const { getByText, getByTestId } = renderComponent();
+    expect(getByTestId('send_amount').props.value).toBe('');
+    fireEvent.press(getByText('Max'));
+    expect(getByTestId('send_amount').props.value).toBe('0.9999685');
+    expect(getByText('$ 3889.87')).toBeTruthy();
   });
 
-  it('displays fiat value for the amount entered', async () => {
-    jest.spyOn(ConversionUtils, 'useConversions').mockReturnValue({
-      getFiatDisplayValue: () => '$ 1200.00',
-      getFiatValue: () => 0,
-      getNativeDisplayValue: () => '',
-      getNativeValue: () => '',
-    });
-    const { getByTestId, getByText } = renderComponent();
-    fireEvent.changeText(getByTestId('send_amount'), '123');
-    expect(getByText('$ 1200.00')).toBeDefined();
+  it('display fiat conversion of amount entered', async () => {
+    (useRoute as jest.MockedFn<typeof useRoute>).mockReturnValue({
+      params: {
+        asset: {
+          name: 'Ethereum',
+          address: TOKEN_ADDRESS_MOCK_1,
+          isNative: true,
+          chainId: '0x1',
+          symbol: 'ETH',
+        },
+      },
+    } as RouteProp<ParamListBase, string>);
+
+    const { getByText, getByTestId } = renderComponent();
+    fireEvent.changeText(getByTestId('send_amount'), '1');
+    expect(getByText('$ 3890')).toBeTruthy();
   });
 
-  it('displays native value for the amount entered if fiat_mode is enabled', async () => {
-    jest.spyOn(ConversionUtils, 'useConversions').mockReturnValue({
-      getFiatDisplayValue: () => '',
-      getFiatValue: () => 0,
-      getNativeDisplayValue: () => 'ETH 0.001',
-      getNativeValue: () => '',
-    });
-    const { getByTestId, getByText } = renderComponent();
+  it('if fiatmode is enabled display native conversion of amount entered', async () => {
+    (useRoute as jest.MockedFn<typeof useRoute>).mockReturnValue({
+      params: {
+        asset: {
+          name: 'Ethereum',
+          address: TOKEN_ADDRESS_MOCK_1,
+          isNative: true,
+          chainId: '0x1',
+          symbol: 'ETH',
+        },
+      },
+    } as RouteProp<ParamListBase, string>);
+
+    const { getByText, getByTestId } = renderComponent();
     fireEvent.press(getByTestId('fiat_toggle'));
-    fireEvent.changeText(getByTestId('send_amount'), '123');
-    expect(getByText('ETH 0.001')).toBeDefined();
+    fireEvent.changeText(getByTestId('send_amount'), '7780');
+    expect(getByText('ETH 2')).toBeTruthy();
+  });
+
+  it('pressing Max in fiat mode should work as expected', () => {
+    (useRoute as jest.MockedFn<typeof useRoute>).mockReturnValue({
+      params: {
+        asset: {
+          name: 'Ethereum',
+          address: TOKEN_ADDRESS_MOCK_1,
+          isNative: true,
+          chainId: '0x1',
+          symbol: 'ETH',
+        },
+      },
+    } as RouteProp<ParamListBase, string>);
+
+    const { getByText, getByTestId } = renderComponent();
+    expect(getByTestId('send_amount').props.value).toBe('');
+    fireEvent.press(getByTestId('fiat_toggle'));
+    fireEvent.press(getByText('Max'));
+    expect(getByTestId('send_amount').props.value).toBe('3889.87746');
+    expect(getByText('ETH 0.99997')).toBeTruthy();
+  });
+
+  it('navigate to next page when continue button is clicked', () => {
+    const { getByText, getByTestId } = renderComponent();
+    fireEvent.changeText(getByTestId('send_amount'), '100');
+    fireEvent.press(getByText('Continue'));
+    expect(mockNavigate).toHaveBeenCalled();
   });
 });
