+++ app/components/Views/confirmations/components/send/asset/asset.test.tsx
@@ -5,9 +5,11 @@ import { backgroundState } from '../../../../../../util/test/initial-root-state'
 import { SendContextProvider } from '../../../context/send-context';
 import { Asset } from './asset';
 
+const mockNavigate = jest.fn();
 jest.mock('@react-navigation/native', () => ({
   ...jest.requireActual('@react-navigation/native'),
   useNavigation: () => ({
+    navigate: mockNavigate,
     goBack: jest.fn(),
   }),
   useRoute: jest.fn().mockReturnValue({
@@ -37,4 +39,11 @@ describe('Asset', () => {
 
     expect(getByText('Asset: NA')).toBeTruthy();
   });
+
+  // to be uncommented after asset selection is implemented
+  // it('navigate to next page when continue button is clicked', () => {
+  //   const { getByText } = renderComponent();
+  //   fireEvent.press(getByText('Continue'));
+  //   expect(mockNavigate).toHaveBeenCalled();
+  // });
 });
