--- app/components/Views/confirmations/components/send/send-root/index.ts
@@ -1,1 +1,0 @@
-export { SendRoot } from './send-root';
