--- app/components/Views/confirmations/components/send/send-root/send-root.styles.ts
@@ -1,8 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styleSheet = () =>
-  StyleSheet.create({
-    container: {
-      padding: 20,
-    },
-  });
