+++ app/components/Views/confirmations/components/send/send-to/send-to.tsx
@@ -1,11 +1,16 @@
 import React from 'react';
 import { View } from 'react-native';
 
+import Button, {
+  ButtonSize,
+  ButtonVariants,
+} from '../../../../../../component-library/components/Buttons/Button';
 import Input from '../../../../../../component-library/components/Form/TextField/foundation/Input';
 import Text, {
   TextColor,
 } from '../../../../../../component-library/components/Texts/Text';
 import { useStyles } from '../../../../../hooks/useStyles';
+import { useSendActions } from '../../../hooks/send/useSendActions';
 import { useSendContext } from '../../../context/send-context';
 import { useToAddressValidation } from '../../../hooks/send/useToAddressValidation';
 import { styleSheet } from './send-to.styles';
@@ -14,6 +19,7 @@ export const SendTo = () => {
   const { styles } = useStyles(styleSheet, {});
   const { to, updateTo } = useSendContext();
   const { toAddressError, toAddressWarning } = useToAddressValidation();
+  const { handleSubmitPress } = useSendActions();
 
   return (
     <View>
@@ -28,6 +34,13 @@ export const SendTo = () => {
       {toAddressWarning && (
         <Text color={TextColor.Warning}>{toAddressWarning}</Text>
       )}
+      <Button
+        label="Continue"
+        disabled={!to || Boolean(toAddressError)}
+        onPress={handleSubmitPress}
+        variant={ButtonVariants.Primary}
+        size={ButtonSize.Lg}
+      />
     </View>
   );
 };
